% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_vi.R
\name{calculate_vi}
\alias{calculate_vi}
\title{Calculate visibility index at points}
\usage{
calculate_vi(dtm, dsm, pts, vi_type, vi_dist, vi_fov = 180, vi_azi = 0)
}
\arguments{
\item{dtm}{SpatRaster digital terrain model at finest resolution available}

\item{dsm}{SpatRaster digital surface model at finest resolution available}

\item{pts}{A dataframe of the points (with an 'x' and 'y' column) for which vi will be calculated}

\item{vi_type}{which type of VI calculation would you like to conduct? One of: omnidir (omnidirectional, 360 degree), directional_single (a single specified view direction), directional_random (multiple random view directions will be sampled)}

\item{vi_dist}{distance to which vi will be calculated (can be a list of distances if you want to look at multiple (e.g. c(100,200,300)))}

\item{vi_fov}{optional argument, with a default value of 180, if vi_type "directional_single" or "directional_random" are selected. Specifies the field of view for the vi calculation}

\item{vi_azi}{optional argument, with a default value of 0, if vi_type "directional_single" specify the direction of vi in degrees values between 0 and 359 (e.g. North = 0, South = 180, East = 90, West = 270)}
}
\value{
dataframe with columns 'x', 'y', 'vi_x' for x in vi_dist, and 'azimuth' if vi_type = 'directional_random'
}
\description{
Uses both a dtm and dsm to calculate visibility index (visible area/ total area) at points.
}
\examples{
# get your dtm and dsm
dsm <- rast("dsm.tif")
dtm <- rast("dtm.tif")

# get your points
my_points <- generate_pts(dtm, dsm, 100, 1000)

# calculate vi
df <- calculate_vi(dtm, dsm, my_points, "omnidir", c(500, 1000))
}
