% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_preds.R
\name{gen_preds}
\alias{gen_preds}
\title{Generate predictors for modeling visibility index}
\usage{
gen_preds(
  dtm,
  dsm,
  pts,
  vi_type,
  vi_fov = 180,
  vi_azi = 0,
  save = TRUE,
  save_dir = getwd()
)
}
\arguments{
\item{dtm}{SpatRaster digital terrain model at finest resolution available}

\item{dsm}{SpatRaster digital surface model at finest resolution available}

\item{pts}{dataframe of your points, must include x, y columns, may include vi columns}

\item{vi_type}{which type of VI calculation did you do? One of: 'omnidir' (omnidirectional, 360 degree), 'directional_single' (a single specified view direction), 'directional_random' (multiple random view directions will be sampled)}

\item{vi_fov}{optional argument, with a default value of 180, if vi_type "directional_single" or "directional_random" are selected. Specifies the field of view for the vi calculation}

\item{vi_azi}{optional argument, with a default value of 0, if vi_type "directional_single" specify the direction of vi in degrees values between 0 and 359 (e.g. North = 0, South = 180, East = 90, West = 270)}

\item{save}{Default is TRUE. If TRUE and if 'omnidir' or 'single_directional' are used then the multiband raster of predictors will also be saved.}

\item{save_dir}{if save = TRUE specify the directory where the multiband raster of predictors (ONLY if using 'omnidir' or 'directional_single') should be saved, default is your working directory.}
}
\value{
dataframe with predictor columns. If save = TRUE a .tif will also be written to file.
}
\description{
Extracts predictor values at points. Predictors include: canopy cover, canopy height, elevation, slope, slope_derivative, curvature (plan and profile), aspect cosine, aspect sine, slope aspect cosine, slope aspect sine, as well as mean and standard deviation predictors  - for each of the the previous predictors - calculated within a focal area with radius x for x in (2, 4, 6, 8, 16, 32)
}
\examples{
# get your dtm and dsm
dsm <- rast("dsm.tif")
dtm <- rast("dtm.tif")

# get your points
my_points <- generate_pts(dtm, dsm, 100, 1000)

# generate predictors
preds <- gen_preds(dtm, dsm, my_points, "omnidir", save = FALSE)
}
