% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_vi.R
\name{map_vi}
\alias{map_vi}
\title{Map VI across a landscape}
\usage{
map_vi(mod, predictors, percent_threads)
}
\arguments{
\item{mod}{random forest model}

\item{predictors}{a multiband raster of all predictors}

\item{percent_threads}{percent of threads that should be used in prediction, default is 50.}
}
\description{
This function takes in a model and a stack of predictor rasters and models VI across each pixel.
}
\examples{

# get your dtm and dsm
dsm <- rast("dsm.tif")
dtm <- rast("dtm.tif")

# get your points
my_points <- generate_pts(dtm, dsm, 100, 1000)

# calculate vi
df <- calculate_vi(dtm, dsm, my_points, "omnidir", 500)

# generate predictors
df_preds <- gen_preds(dtm, dsm, df, "omnidir", save = TRUE)

# model
mod <- model_vi(df_preds, TRUE, FALSE, 80)

rf <- mod[[2]]
predictors <- rast("predictor_raster_stack.tif"))

mymap <- map_vi(rf, predictors, 50)
plot(mymap)
}
