% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_vi.R
\name{model_vi}
\alias{model_vi}
\title{Model VI using a random forest}
\usage{
model_vi(df, cross_validate = FALSE, tune = FALSE, percent_cores = 50)
}
\arguments{
\item{df}{dataframe with predictor variables, target vi values, and an 'x' and 'y' column with coordinates}

\item{cross_validate}{default is FALSE. True will execute a 4 fold spatial cross validation. False will train on all available points. If true, the predictions that are returned are predictions for folds made by models build excluding the respective fold. For example, predictions made on fold one are generated from models built using folds 2, 3, and 4.}

\item{tune}{default is FALSE. TRUE will tune a full model and use tuned hyperparameters (mtry, minimum node size, sample fraction) for modeling.}

\item{percent_cores}{The percent of cores you'd like to use for modeling and tuning (default is 50).}
}
\value{
If cross_validate is FALSE, returns a random forest. If cross_validate is TRUE, returns a list where [\link{1}] is a ataframe with predictions column indicating predictions made by the model and [\link{2}] is the trained random forest model.
}
\description{
Models VI froma dataframe that includes predictor values.
}
\examples{
# get your dtm and dsm
dsm <- rast("dsm.tif")
dtm <- rast("dtm.tif")

# get your points
my_points <- generate_pts(dtm, dsm, 100, 1000)

# calculate vi
df <- calculate_vi(dtm, dsm, my_points, "omnidir", 500)

# generate predictors
df_preds <- gen_preds(dtm, dsm, df, "omnidir", save = FALSE)

# merge dataframes so predictors and vi values are together
df_all <- merge(df, df_preds, by=c("x", "y"))

# model
mod <- model_vi(df_all, TRUE, FALSE, 80)

# print the aggregate model results
print(mod[[2]])
}
