% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VisiMod.R
\name{VisiMod}
\alias{VisiMod}
\title{Model and Map VI}
\usage{
VisiMod(dtm, dsm, num_pts, dist, vi_type, vi_fov, vi_azi, save_dir)
}
\arguments{
\item{dtm}{SpatRaster. A digital terrain model at finest resolution available}

\item{dsm}{SpatRaster. A digital surface model at finest resolution available}

\item{num_pts}{Numeric. Number of points used to train the model}

\item{dist}{Numeric. distance to which VI is calculated. If multiple distances should be considered use c().}

\item{vi_type}{Character. Defines which type of VI calculation you would like to conduct. One of: "omnidir" (omnidirectional, 360 degree), "directional_single" (a single specified view direction for each point), "directional_random" (viewing directions will be randomly assigned to each point).}

\item{vi_fov}{Numeric. Defines the angular field of view, in degrees, of the directional wedge used for the VI calculation. Only used if vi_type == "directional_single". Values must be > 0 and < 360}

\item{vi_azi}{Numeric. Defines the azimuth, or central viewing direction, in degrees, of the directional wedge used for VI calculation.  Only used if vi_type == "directional_single". Values must be >= 0 and < 360.}

\item{save_dir}{Character. The directory where intermediary files will be saved.}
}
\value{
A SpatRaster, or list of SpatRasters, of mapped VI with a potential range of 0-1.
}
\description{
Full workflow for mapping VI using random forest modeling.
}
\examples{
# get your dtm and dsm
dsm <- rast("dsm.tif")
dtm <- rast("dtm.tif")

# model and map VI
outras <- modMapVI(dtm, dsm, 1000, 500, "directional_single", 120, c(0,120,240),  "C:\\\\proj1\\\\")

# create a 3 band raster to display in RGB
stack1 <- terra::`add<-`(outras[[1]], outras[[2]])
multiband_ras <- terra::`add<-`(stack1, outras[[3]])
writeRaster(multiband_ras, "C:\\\\proj1\\\\vi_map.tif")
}
