% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wedge.R
\name{wedge}
\alias{wedge}
\title{Generate wedge-shaped polygon}
\usage{
wedge(x, y, rad, az, theta)
}
\arguments{
\item{x}{Numeric. The x coordinate of the wedge origin point.}

\item{y}{Numeric. The y coordinate of the wedge origin point.}

\item{rad}{Numeric. The radius of the wedge.}

\item{az}{Numeric. The azimuth of the wedge (direction in which the wedge will be drawn) in degrees. Must be between 0 and 359.}

\item{theta}{Numeric. The angular field of view of the wedge in degrees, must be between 1 and 360.}
}
\value{
SpatVector representation of a wedge.
}
\description{
Function that draws a wedge (angular slice of a circle from a central origin point) with user-defined radius, azimuth (central viewing direction), and theta (angular field of view). This is a utility function within the VisiMod library that is used in several functions to aid in directional visibility estimation.
}
\examples{
# draw a east-facing wedge
w <- wedge(0,0, 500, 90, 90)

# plot it
plot(w)
}
