% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_pts.R
\name{gen_pts}
\alias{gen_pts}
\title{Create random points for VI modeling}
\usage{
gen_pts(dtm, dsm, num_pts = 200, max_vis_dist)
}
\arguments{
\item{dtm}{SpatRaster. Digital terrain model at finest resolution available}

\item{dsm}{SpatRaster. Digital surface model at finest resolution available}

\item{num_pts}{Numeric. The number of points generated}

\item{max_vis_dist}{Numeric. The maximum distance (in meters) to which VI will be calculated, to account for study area boundary}
}
\value{
data.frame with columns 'x' and 'y'
}
\description{
Function that generates a user-defined number of randomly distributed points within a study area at an appropriate distance from the study area boundary. Study area extent defined by input digital terrain model (DTM) and digital surface model (DSM). These points are intended for use as training and validation points in the VisiMod visibility modeling workflow.
}
\details{
\itemize{
\item \code{dtm} and \code{dsm} SpatRasters can be defined using the terra library. They should have the same coordinate system, resolution, extent, and origin.
\item The more points generated, the more robust the modeling procedure will be; however, more points will also increase total processing time for subsequent functions in the workflow. The default is set to 200, which should achieve a nice balance between model performance and processing time. We do not recommend generating more than 1000 points.
\item The maximum VI distance will also affect processing time quite dramatically. As distance increases, processing time will increase exponentially. We do not recommend attempting this workflow at distances beyond 3000 m.
}
}
\examples{
# read in dtm and dsm
dtm <- rast("dtm.tif")
dsm <- rast("dsm.tif")

# generate 1,000 points with a buffer of 1,000 meters
generate_pts(dtm, dsm, 1000, 1000)
}
