% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_pts.R
\name{generate_pts}
\alias{generate_pts}
\title{Function that generates a desired number of points at an appropriate distance from study area boundary}
\usage{
generate_pts(dtm, dsm, num_pts, max_vis_dist)
}
\arguments{
\item{dtm}{SpatRaster digital terrain model at finest resolution available}

\item{dsm}{SpatRaster digital surface model at finest resolution available}

\item{num_pts}{The points for which vi will be calculated}

\item{max_vis_dist}{The maximum distance to which VI will be calculated, to account for study area borders}
}
\description{
Function that generates a desired number of points at an appropriate distance from study area boundary
}
\examples{
# generate_pts(dtm, dsm, 1000, 1000)
}
