% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wedge.R
\name{wedge}
\alias{wedge}
\title{Generate wedge-shaped polygon}
\usage{
wedge(x, y, rad, az, theta)
}
\arguments{
\item{x}{the x coordinate of the wedge origin point}

\item{y}{the y coordinate of the wedge origin point}

\item{rad}{the radius of the wedge}

\item{az}{the azimuth of the wedge (direction in which the wedge will be drawn) in degrees. Must be between 0 and 359.}

\item{theta}{the opening angle of the wedge in degrees, must be between 1 and 360}
}
\value{
SpatVector
}
\description{
Function that draws a wedge with user-defined radius, azimuth, and theta.
}
\examples{
# draw a east-facing wedge
w <- wedge(0,0, 500, 90, 90)

# plot it
plot(w)
}
